requirements.txt
Django==5.2.2
djangorestframework==3.16.0
simplejwt==2.0.1
